refactor(parsers): move buildSchemaFromConfigType from utils to parsers-common
To avoid circular dependencies,
functions in utils file should not import functions in parsers-commons.
So, this function should go to parsers-commons file
