+++ packages/react-native-codegen/src/parsers/__tests__/utils-test.js
@@ -11,15 +11,11 @@
 
 'use strict';
 
-const {MockedParser} = require('../parserMock');
-
 const {
   extractNativeModuleName,
   createParserErrorCapturer,
   verifyPlatforms,
   visit,
-  buildSchemaFromConfigType,
-  isModuleRegistryCall,
 } = require('../utils.js');
 const {ParserError} = require('../errors');
 
@@ -294,300 +290,6 @@ describe('visit', () => {
       visit(astNode, visitor);
 
       expect(visitorFunction).toHaveBeenCalledTimes(2);
-    });
-  });
-});
-
-describe('buildSchemaFromConfigType', () => {
-  const parser = new MockedParser();
-
-  const astMock = {
-    type: 'Program',
-    loc: {
-      source: null,
-      start: {line: 2, column: 10},
-      end: {line: 16, column: 62},
-    },
-    range: [11, 373],
-    body: [],
-    comments: [],
-    errors: [],
-  };
-
-  const componentSchemaMock = {
-    filename: 'filename',
-    componentName: 'componentName',
-    extendsProps: [],
-    events: [],
-    props: [],
-    commands: [],
-  };
-
-  const moduleSchemaMock = {
-    type: 'NativeModule',
-    aliases: {},
-    spec: {properties: []},
-    moduleName: '',
-  };
-
-  const wrapComponentSchemaMock = jest.fn();
-  const buildComponentSchemaMock = jest.fn(_ => componentSchemaMock);
-  const wrapModuleSchemaMock = jest.spyOn(
-    require('../parsers-commons'),
-    'wrapModuleSchema',
-  );
-  const buildModuleSchemaMock = jest.fn((_0, _1, _2, _3) => moduleSchemaMock);
-
-  const buildSchemaFromConfigTypeHelper = (
-    configType: 'module' | 'component' | 'none',
-    filename: ?string,
-  ) =>
-    buildSchemaFromConfigType(
-      configType,
-      filename,
-      astMock,
-      wrapComponentSchemaMock,
-      buildComponentSchemaMock,
-      buildModuleSchemaMock,
-      parser,
-    );
-
-  describe('when configType is none', () => {
-    it('returns an empty schema', () => {
-      const schema = buildSchemaFromConfigTypeHelper('none');
-
-      expect(schema).toEqual({modules: {}});
-    });
-  });
-
-  describe('when configType is component', () => {
-    it('calls buildComponentSchema with ast and wrapComponentSchema with the result', () => {
-      buildSchemaFromConfigTypeHelper('component');
-
-      expect(buildComponentSchemaMock).toHaveBeenCalledTimes(1);
-      expect(buildComponentSchemaMock).toHaveBeenCalledWith(astMock);
-      expect(wrapComponentSchemaMock).toHaveBeenCalledTimes(1);
-      expect(wrapComponentSchemaMock).toHaveBeenCalledWith(componentSchemaMock);
-
-      expect(buildModuleSchemaMock).not.toHaveBeenCalled();
-      expect(wrapModuleSchemaMock).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('when configType is module', () => {
-    describe('when filename is undefined', () => {
-      it('throws an error', () => {
-        expect(() => buildSchemaFromConfigTypeHelper('module')).toThrow(
-          'Filepath expected while parasing a module',
-        );
-      });
-    });
-
-    describe('when filename is null', () => {
-      it('throws an error', () => {
-        expect(() => buildSchemaFromConfigTypeHelper('module', null)).toThrow(
-          'Filepath expected while parasing a module',
-        );
-      });
-    });
-
-    describe('when filename is defined and not null', () => {
-      describe('when buildModuleSchema throws', () => {
-        it('throws the error', () => {
-          const parserError = new ParserError(
-            'moduleName',
-            astMock,
-            'Something went wrong',
-          );
-          buildModuleSchemaMock.mockImplementationOnce(() => {
-            throw parserError;
-          });
-
-          expect(() =>
-            buildSchemaFromConfigTypeHelper('module', 'filename'),
-          ).toThrow(parserError);
-        });
-      });
-
-      describe('when buildModuleSchema returns null', () => {
-        it('throws an error', () => {
-          // $FlowIgnore[incompatible-call] - This is to test an invariant
-          buildModuleSchemaMock.mockReturnValueOnce(null);
-
-          expect(() =>
-            buildSchemaFromConfigTypeHelper('module', 'filename'),
-          ).toThrow(
-            'When there are no parsing errors, the schema should not be null',
-          );
-        });
-      });
-
-      describe('when buildModuleSchema returns a schema', () => {
-        it('calls buildModuleSchema with ast and wrapModuleSchema with the result', () => {
-          buildSchemaFromConfigTypeHelper('module', 'filename');
-
-          expect(buildModuleSchemaMock).toHaveBeenCalledTimes(1);
-          expect(buildModuleSchemaMock).toHaveBeenCalledWith(
-            'filename',
-            astMock,
-            expect.any(Function),
-            parser,
-          );
-          expect(wrapModuleSchemaMock).toHaveBeenCalledTimes(1);
-          expect(wrapModuleSchemaMock).toHaveBeenCalledWith(
-            moduleSchemaMock,
-            'filename',
-          );
-
-          expect(buildComponentSchemaMock).not.toHaveBeenCalled();
-          expect(wrapComponentSchemaMock).not.toHaveBeenCalled();
-        });
-      });
-    });
-  });
-
-  describe('isModuleRegistryCall', () => {
-    describe('when node is not of CallExpression type', () => {
-      it('returns false', () => {
-        const node = {
-          type: 'NotCallExpression',
-        };
-        expect(isModuleRegistryCall(node)).toBe(false);
-      });
-    });
-
-    describe('when node is of CallExpressionType', () => {
-      describe('when callee type is not of MemberExpression type', () => {
-        it('returns false', () => {
-          const node = {
-            type: 'CallExpression',
-            callee: {
-              type: 'NotMemberExpression',
-            },
-          };
-          expect(isModuleRegistryCall(node)).toBe(false);
-        });
-      });
-
-      describe('when callee type is of MemberExpression type', () => {
-        describe('when memberExpression has an object of type different than "Identifier"', () => {
-          it('returns false', () => {
-            const node = {
-              type: 'CallExpression',
-              callee: {
-                type: 'MemberExpression',
-                object: {
-                  type: 'NotIdentifier',
-                  name: 'TurboModuleRegistry',
-                },
-              },
-            };
-            expect(isModuleRegistryCall(node)).toBe(false);
-          });
-        });
-
-        describe('when memberExpression has an object of name different than "TurboModuleRegistry"', () => {
-          it('returns false', () => {
-            const node = {
-              type: 'CallExpression',
-              callee: {
-                type: 'MemberExpression',
-                object: {
-                  type: 'Identifier',
-                  name: 'NotTurboModuleRegistry',
-                },
-              },
-            };
-            expect(isModuleRegistryCall(node)).toBe(false);
-          });
-        });
-
-        describe('when memberExpression has an object of type "Identifier" and name "TurboModuleRegistry', () => {
-          describe('when memberExpression has a property of type different than "Identifier"', () => {
-            it('returns false', () => {
-              const node = {
-                type: 'CallExpression',
-                callee: {
-                  type: 'MemberExpression',
-                  object: {
-                    type: 'Identifier',
-                    name: 'TurboModuleRegistry',
-                  },
-                  property: {
-                    type: 'NotIdentifier',
-                    name: 'get',
-                  },
-                },
-              };
-              expect(isModuleRegistryCall(node)).toBe(false);
-            });
-          });
-
-          describe('when memberExpression has a property of name different than "get" or "getEnforcing', () => {
-            it('returns false', () => {
-              const node = {
-                type: 'CallExpression',
-                callee: {
-                  type: 'MemberExpression',
-                  object: {
-                    type: 'Identifier',
-                    name: 'TurboModuleRegistry',
-                  },
-                  property: {
-                    type: 'Identifier',
-                    name: 'NotGet',
-                  },
-                },
-              };
-              expect(isModuleRegistryCall(node)).toBe(false);
-            });
-          });
-
-          describe('when memberExpression has a property of type "Identifier" and of name "get" or "getEnforcing', () => {
-            describe('when memberExpression is computed', () => {
-              it('returns false', () => {
-                const node = {
-                  type: 'CallExpression',
-                  callee: {
-                    type: 'MemberExpression',
-                    object: {
-                      type: 'Identifier',
-                      name: 'TurboModuleRegistry',
-                    },
-                    property: {
-                      type: 'Identifier',
-                      name: 'get',
-                    },
-                    computed: true,
-                  },
-                };
-                expect(isModuleRegistryCall(node)).toBe(false);
-              });
-            });
-
-            describe('when memberExpression is not computed', () => {
-              it('returns true', () => {
-                const node = {
-                  type: 'CallExpression',
-                  callee: {
-                    type: 'MemberExpression',
-                    object: {
-                      type: 'Identifier',
-                      name: 'TurboModuleRegistry',
-                    },
-                    property: {
-                      type: 'Identifier',
-                      name: 'get',
-                    },
-                    computed: false,
-                  },
-                };
-                expect(isModuleRegistryCall(node)).toBe(true);
-              });
-            });
-          });
-        });
-      });
     });
   });
 });
